Refactor command line handlers: Remove unused WiFi and BLE command functions
- Removed various WiFi scan, attack, and beacon spam command handlers.
- Eliminated BLE command handlers related to scanning and wardriving.
- Cleaned up unused function declarations and definitions for better maintainability.
- Added includes for WiFi and BLE CLI handlers to support future functionality.
