+++ include/core/cli-handlers/ble-cli-handler.h
@@ -1,0 +1,12 @@
+#pragma once
+
+void handle_ble_scan_cmd(int argc, char **argv);
+void handle_ble_spam_cmd(int argc, char **argv);
+void handle_list_airtags_cmd(int argc, char **argv);
+void handle_select_airtag(int argc, char **argv);
+void handle_spoof_airtag(int argc, char **argv);
+void handle_stop_spoof(int argc, char **argv);
+void handle_ble_wardriving(int argc, char **argv);
+void handle_list_flippers_cmd(int argc, char **argv);
+void handle_select_flipper_cmd(int argc, char **argv);
+void handle_stop_flipper(int argc, char **argv);
